refactor(tools): split pipeline_init_with_context.go into focused sub-files (Phase 3b)
Extracted context-parsing helpers into context_fetcher.go and workspace
initialisation helpers into workspace_init.go; moved deriveSpecName from
pipeline_init.go to workspace_init.go; pipeline_init_with_context.go
retains only the handler, 3-call flow functions, flag/confirmation types,
and their parsers.
