+++ mcp-server/internal/tools/pipeline_init_with_context.go
@@ -19,8 +19,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"os"
-	"path/filepath"
 	"slices"
 	"strings"
 
@@ -62,25 +60,6 @@ type UserConfirmationPrompt struct {
 	IsMainBranch        bool                                `json:"is_main_branch"`
 	Message             string                              `json:"message"`
 	EnrichedRequestBody string                              `json:"enriched_request_body,omitempty"`
-}
-
-// externalContext holds parsed GitHub/Jira context fields.
-type externalContext struct {
-	// Source identifiers from pipeline_init result — used in request.md front matter.
-	SourceURL string
-	SourceID  string
-	// TaskText is the original task text for text source type pipelines.
-	// Populated from the top-level task_text MCP parameter (not from external_context map).
-	TaskText string
-	// GitHub fields
-	GitHubLabels []string
-	GitHubTitle  string
-	GitHubBody   string
-	// Jira fields
-	JiraIssueType   string
-	JiraStoryPoints int
-	JiraSummary     string
-	JiraDescription string
 }
 
 // pipelineFlags holds parsed flag fields from the flags parameter.
@@ -320,139 +299,7 @@ func handleSecondCall(
 	})
 }
 
-// initWorkspace executes the 8-step I/O sequence for the second call.
-// It creates the workspace directory, initialises state, applies all configuration in
-// a single write via sm.Configure, and writes request.md.
-// enrichedBody is the pre-built request.md body from the discussion call; "" means use
-// buildRequestMDWithBody defaults (extCtx.TaskText for text source, GitHub/Jira content otherwise).
-// Returns the request.md content on success.
-func initWorkspace(
-	sm *state.StateManager,
-	workspace, specName string,
-	flags pipelineFlags,
-	uc userConfirmation,
-	branchName string,
-	flowTemplate string,
-	skippedPhases []string,
-	extCtx externalContext,
-	enrichedBody string,
-) (string, error) {
-	// Validate workspace path — reject non-ASCII characters so that
-	// multibyte input (e.g. Japanese) never produces an unreadable directory name.
-	if hasNonASCII(workspace) {
-		return "", fmt.Errorf("workspace path %q contains non-ASCII characters; use only ASCII in directory names", workspace)
-	}
-
-	// Create workspace directory.
-	if err := os.MkdirAll(workspace, 0o750); err != nil {
-		return "", fmt.Errorf("MkdirAll %q: %w", workspace, err)
-	}
-
-	// Check state.json doesn't already exist.
-	stateFile := filepath.Join(workspace, "state.json")
-	if _, err := os.Stat(stateFile); err == nil {
-		return "", fmt.Errorf("workspace %q already initialised: state.json exists", workspace)
-	}
-
-	// sm.Init.
-	if err := sm.Init(workspace, specName); err != nil {
-		return "", fmt.Errorf("sm.Init: %w", err)
-	}
-
-	// Apply all configuration in a single write to state.json.
-	// Branch name is pre-derived by the caller (either current branch or
-	// DeriveBranchName), avoiding a separate SetBranch call.
-	cfg := state.PipelineConfig{
-		Effort:           uc.Effort,
-		FlowTemplate:     flowTemplate,
-		AutoApprove:      flags.Auto,
-		SkipPR:           flags.SkipPR,
-		Debug:            flags.Debug,
-		SkippedPhases:    skippedPhases,
-		UseCurrentBranch: uc.UseCurrentBranch,
-		Branch:           branchName,
-	}
-	if err := sm.Configure(workspace, cfg); err != nil {
-		return "", fmt.Errorf("configure: %w", err)
-	}
-
-	// Write request.md.
-	body := enrichedBody
-	if body == "" {
-		body = extCtx.TaskText
-	}
-	requestMD := buildRequestMDWithBody(extCtx, body)
-	reqPath := filepath.Join(workspace, "request.md")
-	if err := os.WriteFile(reqPath, []byte(requestMD), 0o600); err != nil {
-		return "", fmt.Errorf("write request.md: %w", err)
-	}
-
-	return requestMD, nil
-}
-
-// ---------- helpers ----------
-
-// parseExternalContext extracts GitHub/Jira context fields from the args map.
-func parseExternalContext(args map[string]any) (externalContext, error) {
-	var extCtx externalContext
-
-	raw, ok := args["external_context"]
-	if !ok || raw == nil {
-		return extCtx, nil
-	}
-
-	// Round-trip through JSON to normalize types.
-	data, err := json.Marshal(raw)
-	if err != nil {
-		return extCtx, fmt.Errorf("marshal external_context: %w", err)
-	}
-	var m map[string]any
-	if err := json.Unmarshal(data, &m); err != nil {
-		return extCtx, fmt.Errorf("unmarshal external_context: %w", err)
-	}
-
-	extCtx.SourceURL = stringField(m, "source_url")
-	extCtx.SourceID = stringField(m, "source_id")
-	extCtx.GitHubTitle = stringField(m, "github_title")
-	extCtx.GitHubBody = stringField(m, "github_body")
-	extCtx.JiraIssueType = stringFieldAlt(m, "jira_issue_type", "issue_type")
-	extCtx.JiraSummary = stringFieldAlt(m, "jira_summary", "summary")
-	extCtx.JiraDescription = stringFieldAlt(m, "jira_description", "description")
-
-	// Parse github_labels (array of strings).
-	if labelsRaw, ok := m["github_labels"]; ok {
-		switch v := labelsRaw.(type) {
-		case []any:
-			for _, l := range v {
-				if s, ok := l.(string); ok {
-					extCtx.GitHubLabels = append(extCtx.GitHubLabels, s)
-				}
-			}
-		case []string:
-			extCtx.GitHubLabels = v
-		}
-	}
-
-	// Parse jira_story_points (number), with "story_points" as fallback alias.
-	if _, ok := m["jira_story_points"]; !ok {
-		if sp, ok2 := m["story_points"]; ok2 {
-			m["jira_story_points"] = sp
-		}
-	}
-	if spRaw, ok := m["jira_story_points"]; ok {
-		switch v := spRaw.(type) {
-		case float64:
-			extCtx.JiraStoryPoints = int(v)
-		case int:
-			extCtx.JiraStoryPoints = v
-		case json.Number:
-			n, _ := v.Int64()
-			extCtx.JiraStoryPoints = int(n)
-		}
-	}
-
-	return extCtx, nil
-}
+// ---------- flag and confirmation parsers ----------
 
 // parseFlags extracts flag fields from the args map.
 func parseFlags(args map[string]any) (pipelineFlags, error) {
@@ -500,118 +347,4 @@ func parseUserConfirmation(raw any) (userConfirmation, error) {
 	uc.UseCurrentBranch = boolField(m, "use_current_branch")
 	uc.EnrichedRequestBody = stringField(m, "enriched_request_body")
 	return uc, nil
-}
-
-// buildRequestMDWithBody constructs the request.md content.
-// For text source type: uses the body parameter directly (enables both task_text passthrough
-// and discussion-enriched body). For github_issue/jira_issue: ignores body and uses the
-// GitHub/Jira fields as before.
-func buildRequestMDWithBody(extCtx externalContext, body string) string {
-	var sb strings.Builder
-
-	// Determine source_type and body.
-	sourceType := "text"
-	var resolvedBody string
-
-	switch {
-	case extCtx.GitHubTitle != "" || extCtx.GitHubBody != "":
-		sourceType = "github_issue"
-		resolvedBody = strings.TrimSpace(extCtx.GitHubTitle + "\n\n" + extCtx.GitHubBody)
-	case extCtx.JiraIssueType != "" || extCtx.JiraSummary != "" || extCtx.JiraDescription != "":
-		sourceType = "jira_issue"
-		resolvedBody = strings.TrimSpace(extCtx.JiraSummary + "\n\n" + extCtx.JiraDescription)
-	default:
-		// text source: use the body parameter directly.
-		resolvedBody = body
-	}
-
-	sb.WriteString("---\n")
-	sb.WriteString("source_type: ")
-	sb.WriteString(sourceType)
-	sb.WriteString("\n")
-	if extCtx.SourceURL != "" {
-		sb.WriteString("source_url: ")
-		sb.WriteString(extCtx.SourceURL)
-		sb.WriteString("\n")
-	}
-	if extCtx.SourceID != "" {
-		sb.WriteString("source_id: ")
-		sb.WriteString(extCtx.SourceID)
-		sb.WriteString("\n")
-	}
-	sb.WriteString("---\n")
-
-	if resolvedBody != "" {
-		sb.WriteString("\n")
-		sb.WriteString(resolvedBody)
-		sb.WriteString("\n")
-	}
-
-	return sb.String()
-}
-
-// buildEnrichedRequestBody combines the original task text with the discussion answers
-// in a structured markdown format.
-func buildEnrichedRequestBody(taskText, discussionAnswers string) string {
-	return fmt.Sprintf("%s\n\n## Discussion Answers\n\n%s", taskText, discussionAnswers)
-}
-
-// defaultDiscussionQuestions returns the set of generic clarification questions
-// presented to the user when --discuss is active.
-func defaultDiscussionQuestions() []string {
-	return []string{
-		"What is the expected outcome or definition of done for this task?",
-		"Are there any constraints, non-goals, or out-of-scope items?",
-		"Are there any specific implementation details, preferences, or context the agent should know?",
-	}
-}
-
-// ---------- field extraction helpers ----------
-
-func stringField(m map[string]any, key string) string {
-	v, ok := m[key]
-	if !ok || v == nil {
-		return ""
-	}
-	s, _ := v.(string)
-	return s
-}
-
-// stringFieldAlt tries the primary key first, then falls back to the alt key.
-// This allows callers to pass either "jira_summary" or "summary" as the field name.
-func stringFieldAlt(m map[string]any, primary, alt string) string {
-	if s := stringField(m, primary); s != "" {
-		return s
-	}
-	return stringField(m, alt)
-}
-
-func boolField(m map[string]any, key string) bool {
-	v, ok := m[key]
-	if !ok || v == nil {
-		return false
-	}
-	b, _ := v.(bool)
-	return b
-}
-
-// applyWorkspaceSlug replaces the slug portion of a workspace path with the
-// LLM-generated slug. If slugify produces an empty result (e.g. pure Japanese input),
-// the original workspace path is returned unchanged.
-func applyWorkspaceSlug(workspace, rawSlug string) string {
-	cleaned := slugify(rawSlug)
-	if cleaned == "" {
-		return workspace
-	}
-	return replaceWorkspaceSlug(workspace, cleaned)
-}
-
-// hasNonASCII guards workspace paths against unreadable multibyte characters (e.g. Japanese).
-func hasNonASCII(s string) bool {
-	for _, r := range s {
-		if r > 0x7F {
-			return true
-		}
-	}
-	return false
 }
